pyproject.toml
[project]
name = "flea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "alembic>=1.16.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fa-filter>=0.2.1",
    "fastapi>=0.115.13",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[tool.ruff]
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
]
line-length = 88
indent-width = 4
fix = true

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "F", # Pyflakes
    "UP", # pyupgrade
    "B", # flake8-bugbear
    "SIM", # flake8-simplify
    "I", # isort
    "PL", # pylint
]
ignore = [
    "E501", # line too long
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
follow_untyped_imports = true
disallow_untyped_calls = false
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "data",
    "tests",
]

[dependency-groups]
dev = [
    "aiodocker>=0.24.0",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "faker>=37.4.0",
    "httpx>=0.28.1",
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "ruff>=0.12.0",
]
